+++ sample-apps/react/react-video-demo/src/components/MobileAppBanner/index.ts
@@ -1,0 +1,1 @@
+export { MobileAppBanner as default } from './MobileAppBanner';